fix get + user in post
