+++ src/main/java/net/atos/entng/calendar/models/reminders/ReminderFrontEndModel.java
@@ -16,8 +16,8 @@ public class ReminderFrontEndModel implements IModel<ReminderFrontEndModel> {
     public ReminderFrontEndModel (JsonObject reminder) {
         if (reminder.containsKey(Field._ID)) this._id = reminder.getString(Field._ID);
         this.eventId = reminder.getString(Field.EVENTID_CAMEL, "");
-        this.reminderType = new ReminderTypeModel((JsonObject) reminder.getValue(Field.REMINDERTYPE, new ReminderTypeModel(new JsonObject())));
-        this.reminderFrequency = new ReminderFrequencyFrontEndModel((JsonObject) reminder.getValue(Field.REMINDERFREQUENCY, new ReminderFrequencyFrontEndModel(new JsonObject()))); //pass through converter
+        this.reminderType = new ReminderTypeModel((JsonObject) reminder.getValue(Field.REMINDERTYPE, new JsonObject()));
+        this.reminderFrequency = new ReminderFrequencyFrontEndModel((JsonObject) reminder.getValue(Field.REMINDERFREQUENCY, new JsonObject()));
     }
 
     public ReminderTypeModel getReminderType() {
