+++ src/main/java/net/atos/entng/calendar/models/reminders/ReminderModel.java
@@ -5,6 +5,7 @@ import io.vertx.core.json.JsonObject;
 import net.atos.entng.calendar.core.constants.Field;
 import net.atos.entng.calendar.helpers.IModelHelper;
 import net.atos.entng.calendar.models.IModel;
+import net.atos.entng.calendar.models.OwnerModel;
 import net.atos.entng.calendar.models.User;
 
 import java.util.ArrayList;
@@ -15,14 +16,14 @@ public class ReminderModel implements IModel<ReminderModel> {
 
     private String _id;
     private String eventId;
-    private User owner;
+    private OwnerModel owner;
     private ReminderTypeModel reminderType;
     private List<String> reminderFrequency;
 
     public ReminderModel(JsonObject reminder) {
         if (reminder.containsKey(Field._ID))  this._id = reminder.getString(Field._ID);
         this.eventId = reminder.getString(Field.EVENTID_CAMEL, "");
-        this.owner =  new User((JsonObject) reminder.getValue(Field.OWNER, new JsonObject()));
+        this.owner =  new OwnerModel((JsonObject) reminder.getValue(Field.OWNER, new JsonObject()));
         this.reminderType = new ReminderTypeModel((JsonObject) reminder.getValue(Field.REMINDERTYPE, new JsonObject()));
         this.reminderFrequency = reminder.getJsonArray(Field.REMINDERFREQUENCY, new JsonArray()).getList();;
     }
@@ -44,11 +45,11 @@ public class ReminderModel implements IModel<ReminderModel> {
         this.eventId = eventId;
     }
 
-    public User getOwner() {
+    public OwnerModel getOwner() {
         return owner;
     }
 
-    public void setOwner(User owner) {
+    public void setOwner(OwnerModel owner) {
         this.owner = owner;
     }
 
@@ -70,6 +71,8 @@ public class ReminderModel implements IModel<ReminderModel> {
 
     public JsonObject toJson() {
         JsonObject reminderObject = IModelHelper.toJson(this, true, false);
+
+        reminderObject.put(Field.OWNER, this.getOwner().toJson());
 
         return reminderObject;
     }
